Начал делать передачу автопоказаний. Решил что инкремент прибавляться будет к последнему значению по данным сайта

+++ database/database_description.sql
@@ -22,8 +22,8 @@ CREATE TABLE gas_nn_accounts
 CREATE TABLE gas_nn_meter_readings
                    (id integer primary key,
                    account integer NOT NULL,
-                   date integer NOT NULL DEFAULT 0,
+                   date float NOT NULL DEFAULT 0,
                    current_value float NOT NULL DEFAULT 0,
                    is_sent bit NOT NULL DEFAULT 0,
-                   date_of_sending integer NOT NULL DEFAULT 0,
+                   date_of_sending float NOT NULL DEFAULT 0,
                    FOREIGN KEY (account) REFERENCES gas_nn_accounts(id));
